add initial base classes
Create the base class for the application and associated test case. The test case will be used to run the acceptance test and validate a functioning application.

It will be the responsibility of the candidate to provide an implementation of the test case that can be run to validate functionality.